Remove map marker clustering
