+++ client/src/helpers/Constants.js
@@ -1,7 +1,6 @@
 export const MENU_ITEMS = [
   { text: "About Us", link: "/about" },
   { text: "Find Food", link: "/organizations" },
-  { text: "Find Food (Experimental)", link: "/organizationsexp" },
   { text: "Donate", link: "/donate" },
   { text: "FAQs", link: "/faqs" },
   { text: "Suggest New Listing", link: "/suggestion" },
